+++ cli/client/interface.go
@@ -24,6 +24,7 @@ type APIClient interface {
 	SilencedAPIClient
 	GenericClient
 	ClusterMemberClient
+	LicenseClient
 }
 
 // GenericClient exposes generic resource methods.
@@ -34,7 +35,7 @@ type GenericClient interface {
 
 // AuthenticationAPIClient client methods for authenticating
 type AuthenticationAPIClient interface {
-	CreateAccessToken(url string, userid string, secret string) (*types.Tokens, error)
+	CreateAccessToken(url string, userid string, secret string) (*types.Tokens, string, error)
 	Logout(token string) error
 	RefreshAccessToken(refreshToken string) (*types.Tokens, error)
 }
@@ -185,4 +186,24 @@ type SilencedAPIClient interface {
 type ClusterMemberClient interface {
 	// MemberList lists cluster members
 	MemberList() (*clientv3.MemberListResponse, error)
+
+	// MemberAdd adds a cluster member
+	MemberAdd(peerAddrs []string) (*clientv3.MemberAddResponse, error)
+
+	// MemberUpdate updates a cluster member
+	MemberUpdate(id uint64, peerAddrs []string) (*clientv3.MemberUpdateResponse, error)
+
+	// MemberRemove removes a cluster member
+	MemberRemove(id uint64) (*clientv3.MemberRemoveResponse, error)
+}
+
+// LicenseClient specifies the enteprise client methods for license management.
+// This is a temporary workaround until
+// https://github.com/sensu/sensu-go/issues/1870 is implemented
+type LicenseClient interface {
+	// FetchLicense fetches the installed license
+	FetchLicense() (interface{}, error)
+
+	// UpdateLicense updates the installed enterprise license
+	UpdateLicense(license interface{}) error
 }