+++ dashboard/src/components/partials/EventDetailsContainer/EventDetailsResolveAction.js
@@ -1,18 +1,17 @@
 import React from "react";
 import PropTypes from "prop-types";
 import gql from "graphql-tag";
-import Button from "@material-ui/core/Button";
+import { withApollo } from "react-apollo";
 import resolveEvent from "/mutations/resolveEvent";
 
 class EventDetailsResolveAction extends React.PureComponent {
   static propTypes = {
     client: PropTypes.object.isRequired,
-    disabled: PropTypes.bool,
+    children: PropTypes.func.isRequired,
     event: PropTypes.object,
   };
 
   static defaultProps = {
-    disabled: false,
     event: null,
   };
 
@@ -27,47 +26,24 @@ class EventDetailsResolveAction extends React.PureComponent {
     `,
   };
 
-  state = {
-    locked: false,
-  };
-
-  resolveStart() {
-    this.setState({ locked: true });
-  }
-
-  resolveEnd() {
-    this.setState({ locked: false });
-  }
-
   resolveEvent = () => {
-    const {
-      client,
-      event: { id },
-    } = this.props;
-    if (this.state.locked) {
+    const { client, event } = this.props;
+    if (event.check.status === 0) {
       return;
     }
 
-    this.resolveStart();
-    resolveEvent(client, { id }).then(
-      () => this.resolveEnd(),
-      error => {
-        this.resolveEnd();
-        throw error;
-      },
-    );
+    resolveEvent(client, event);
   };
 
   render() {
-    const { disabled: disabledProp, event, ...props } = this.props;
-    const disabled = disabledProp || event.check.status === 0;
+    const canResolve = this.props.event && this.props.event.check.status === 0;
+    const childProps = {
+      canResolve,
+      resolve: this.resolveEvent,
+    };
 
-    return (
-      <Button onClick={this.resolveEvent} disabled={disabled} {...props}>
-        Resolve
-      </Button>
-    );
+    return this.props.children(childProps);
   }
 }
 
-export default EventDetailsResolveAction;
+export default withApollo(EventDetailsResolveAction);